requirements.txt
pandas==1.3.1
streamlit==0.87.0
Keras==2.4.3
numpy==1.18.5
matplotlib==3.4.2
tensorflow_hub==0.12.0
tensorflow==2.3.0
h5py==2.10.0
imageio==2.9.0
Pillow==8.3.1
